requirements.txt
environs
ffmpeg-python
tinytag
pydantic